utils/file_loader.py
import os
from PyPDF2 import PdfReader
import docx

def load_txt(file_path):
    with open(file_path, "r", encoding="utf-8") as f:
        return f.read()

def load_pdf(file_path):
    try:
        reader = PdfReader(file_path)
        text = ""
        for page in reader.pages:
            text += page.extract_text() or ""
        return text
    except Exception as e:
        return f"❌ Error loading PDF: {str(e)}"


def load_docx(file_path):
    try:
        doc = docx.Document(file_path)
        return "\n".join([para.text for para in doc.paragraphs if para.text.strip()])
    except Exception as e:
        return f"❌ Error loading DOCX: {str(e)}"

def load_document(file_path):
    ext = os.path.splitext(file_path)[1].lower()
    print(f"🔍 Loading file: {file_path} (type: {ext})")

    if ext == ".pdf":
        return load_pdf(file_path)
    elif ext == ".docx":
        return load_docx(file_path)
    elif ext == ".txt":
        return load_txt(file_path)
    else:
        raise ValueError(f"Unsupported file type: {ext}")