requirements.txt
# Core dependencies
chromadb
sentence-transformers

# Document parsing
PyPDF2
python-docx

# Optional UI frameworks
gradio
streamlit

# Optional: if you want to use FAISS instead of ChromaDB
faiss-cpu

# Optional: for environment variable handling or config
python-dotenv

# Optional: for progress bars during ingestion
tqdm